+++ server/db/db.go
@@ -7,7 +7,7 @@ import (
 	"github.com/llr104/LiFrame/utils"
 )
 
-func InitDataBase()  {
+func Init()  {
 
 	dbConfig := utils.GlobalObject.AppConfig.DataBase
 	if  dbConfig.Name == "" || dbConfig.User == "" || dbConfig.Password == "" || dbConfig.IP == "" || dbConfig.Port == 0{
