+++ server/gameslg/enter.go
@@ -5,43 +5,13 @@ import (
 	"github.com/llr104/LiFrame/core/liFace"
 	"github.com/llr104/LiFrame/core/liNet"
 	"github.com/llr104/LiFrame/proto"
-	"github.com/llr104/LiFrame/server/gameutils"
 	"github.com/llr104/LiFrame/utils"
 )
-
-var game mainLogic
-
-type mainLogic struct {
-
-}
-
-/*
-返回true:用户离开了游戏，返回false:用户断线，保留用户的游戏状态
-*/
-func (s *mainLogic) UserOffLine(userId uint32) bool{
-	playerMgr.ReleasePlayer(userId)
-	return true
-}
-
-func (s *mainLogic) UserOnLine(userId uint32){
-	utils.Log.Info("UserOnLine: %d", userId)
-}
-
-func (s *mainLogic) UserLogout(userId uint32) bool{
-	return s.UserOffLine(userId)
-}
-
-
-func (s *mainLogic) ShutDown(){
-	utils.Log.Info("ShutDown")
-}
 
 var Enter enterGame
 
 func init() {
 	Enter = enterGame{}
-	game = mainLogic{}
-	gameutils.STS.SetGame(&game)
 }
